+++ include/nanobind/nb_attr.h
@@ -459,11 +459,11 @@ process_postcall(PyObject **args, std::integral_constant<size_t, NArgs>,
 template <size_t NArgs, typename Policy>
 NB_INLINE void
 process_postcall(PyObject **args, std::integral_constant<size_t, NArgs> nargs,
-                 PyObject *result, call_policy<Policy> *) {
+                 PyObject *&result, call_policy<Policy> *) {
     // result_guard avoids leaking a reference to the return object
     // if postcall throws an exception
     object result_guard = steal(result);
-    Policy::postcall(args, nargs, handle(result));
+    Policy::postcall(args, nargs, result);
     result_guard.release();
 }